factor out trimCodeBlock
